Move the definition to a more appropriate place.
