+++ linera-sdk/src/abis/fungible.rs
@@ -71,51 +71,6 @@ pub enum FungibleResponse {
     TickerSymbol(String),
 }
 
-/// A delegated fungible operation
-#[derive(Debug, Deserialize, Serialize, GraphQLMutationRootInCrate)]
-pub enum DelegatedFungibleOperation {
-    /// Approve the transfer of tokens
-    Approve {
-        /// Owner to transfer from
-        owner: AccountOwner,
-        /// The spender account
-        spender: AccountOwner,
-        /// Maximum amount to be transferred
-        allowance: Amount,
-    },
-    /// Transfers tokens from a (locally owned) account to a (possibly remote) account.
-    Transfer {
-        /// Owner to transfer from
-        owner: AccountOwner,
-        /// Amount to be transferred
-        amount: Amount,
-        /// Target account to transfer the amount to
-        target_account: Account,
-    },
-    /// Transfers tokens from a (locally owned) account to a (possibly remote) account by using the allowance.
-    TransferFrom {
-	/// Owner to transfer from
-        owner: AccountOwner,
-        /// The spender of the amount.
-        spender: AccountOwner,
-        /// Amount to be transferred
-        amount: Amount,
-        /// Target account to transfer the amount to
-        target_account: Account,
-    },
-    /// Same as `Transfer` but the source account may be remote. Depending on its
-    /// configuration, the target chain may take time or refuse to process
-    /// the message.
-    Claim {
-        /// Source account to claim amount from
-        source_account: Account,
-        /// Amount to be claimed
-        amount: Amount,
-        /// Target account to claim the amount into
-        target_account: Account,
-    },
-}
-
 /// The initial state to instantiate fungible with
 #[derive(Clone, Debug, Deserialize, Eq, Ord, PartialEq, PartialOrd, Serialize)]
 pub struct InitialState {
